+++ instrumentation/elasticsearch/elasticsearch-transport-5.3/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/elasticsearch/transport/v5_3/Elasticsearch53TransportClientTest.java
@@ -5,60 +5,31 @@
 
 package io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.v5_3;
 
-import static io.opentelemetry.instrumentation.testing.util.TelemetryDataUtil.orderByRootSpanName;
-import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.equalTo;
-import static io.opentelemetry.semconv.ExceptionAttributes.EXCEPTION_TYPE;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.elasticsearch.cluster.ClusterName.CLUSTER_NAME_SETTING;
-import static org.junit.jupiter.api.Named.named;
 
-import io.opentelemetry.api.common.AttributeKey;
-import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.instrumentation.testing.internal.AutoCleanupExtension;
-import io.opentelemetry.instrumentation.testing.util.ThrowingSupplier;
-import io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.AbstractElasticsearchClientTest;
-import io.opentelemetry.sdk.trace.data.StatusData;
-import io.opentelemetry.semconv.NetworkAttributes;
-import io.opentelemetry.semconv.incubating.DbIncubatingAttributes;
+import io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.AbstractElasticsearchTransportClientTest;
 import java.io.File;
 import java.util.Collections;
 import java.util.UUID;
-import java.util.stream.Stream;
 import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest;
-import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
-import org.elasticsearch.action.get.GetResponse;
-import org.elasticsearch.action.index.IndexResponse;
 import org.elasticsearch.client.transport.TransportClient;
-import org.elasticsearch.cluster.health.ClusterHealthStatus;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.transport.TransportAddress;
 import org.elasticsearch.env.Environment;
-import org.elasticsearch.index.IndexNotFoundException;
 import org.elasticsearch.node.InternalSettingsPreparer;
 import org.elasticsearch.node.Node;
 import org.elasticsearch.transport.Netty3Plugin;
-import org.elasticsearch.transport.RemoteTransportException;
 import org.elasticsearch.transport.TransportService;
 import org.elasticsearch.transport.client.PreBuiltTransportClient;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.api.io.TempDir;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class Elasticsearch53TransportClientTest extends AbstractElasticsearchClientTest {
+class Elasticsearch53TransportClientTest extends AbstractElasticsearchTransportClientTest {
   private static final Logger logger =
       LoggerFactory.getLogger(Elasticsearch53TransportClientTest.class);
-
-  @RegisterExtension static final AutoCleanupExtension cleanup = AutoCleanupExtension.create();
 
   private static final String clusterName = UUID.randomUUID().toString();
   private static Node testNode;
@@ -132,244 +103,13 @@ class Elasticsearch53TransportClientTest extends AbstractElasticsearchClientTest
     return client;
   }
 
-  private Stream<Arguments> healthArguments() {
-    return Stream.of(
-        Arguments.of(
-            named(
-                "sync",
-                (ThrowingSupplier<ClusterHealthStatus, Exception>) this::clusterHealthSync)),
-        Arguments.of(
-            named(
-                "async",
-                (ThrowingSupplier<ClusterHealthStatus, Exception>) this::clusterHealthAsync)));
+  @Override
+  protected String getAddress() {
+    return tcpPublishAddress.getAddress();
   }
 
-  @ParameterizedTest
-  @MethodSource("healthArguments")
-  void elasticsearchStatus(ThrowingSupplier<ClusterHealthStatus, Exception> supplier)
-      throws Exception {
-    ClusterHealthStatus clusterHealthStatus = testing.runWithSpan("parent", supplier);
-
-    assertThat(clusterHealthStatus.name()).isEqualTo("GREEN");
-
-    testing.waitAndAssertTraces(
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span -> span.hasName("parent").hasKind(SpanKind.INTERNAL).hasNoParent(),
-                span ->
-                    span.hasName("ClusterHealthAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasParent(trace.getSpan(0))
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_ADDRESS,
-                                tcpPublishAddress.getAddress()),
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_PORT, tcpPublishAddress.getPort()),
-                            equalTo(
-                                DbIncubatingAttributes.DB_SYSTEM,
-                                DbIncubatingAttributes.DbSystemValues.ELASTICSEARCH),
-                            equalTo(DbIncubatingAttributes.DB_OPERATION, "ClusterHealthAction"),
-                            equalTo(ELASTICSEARCH_ACTION, "ClusterHealthAction"),
-                            equalTo(ELASTICSEARCH_REQUEST, "ClusterHealthRequest")),
-                span ->
-                    span.hasName("callback")
-                        .hasKind(SpanKind.INTERNAL)
-                        .hasParent(trace.getSpan(0))));
-  }
-
-  private Stream<Arguments> errorArguments() {
-    return Stream.of(
-        Arguments.of(
-            named("sync", (Runnable) () -> prepareGetSync("invalid-index", "test-type", "1"))),
-        Arguments.of(
-            named("async", (Runnable) () -> prepareGetAsync("invalid-index", "test-type", "1"))));
-  }
-
-  @ParameterizedTest
-  @MethodSource("errorArguments")
-  void elasticsearchError(Runnable action) {
-    assertThatThrownBy(() -> testing.runWithSpan("parent", action::run))
-        .isInstanceOf(IndexNotFoundException.class)
-        .hasMessage("no such index");
-
-    testing.waitAndAssertTraces(
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
-                    span.hasName("parent")
-                        .hasKind(SpanKind.INTERNAL)
-                        .hasNoParent()
-                        .hasStatus(StatusData.error())
-                        .hasException(new IndexNotFoundException("no such index")),
-                span ->
-                    span.hasName("GetAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasParent(trace.getSpan(0))
-                        .hasStatus(StatusData.error())
-                        .hasEventsSatisfyingExactly(
-                            event ->
-                                event
-                                    .hasName("exception")
-                                    .hasAttributesSatisfying(
-                                        equalTo(
-                                            EXCEPTION_TYPE,
-                                            RemoteTransportException.class.getName())))
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(
-                                DbIncubatingAttributes.DB_SYSTEM,
-                                DbIncubatingAttributes.DbSystemValues.ELASTICSEARCH),
-                            equalTo(DbIncubatingAttributes.DB_OPERATION, "GetAction"),
-                            equalTo(ELASTICSEARCH_ACTION, "GetAction"),
-                            equalTo(ELASTICSEARCH_REQUEST, "GetRequest"),
-                            equalTo(ELASTICSEARCH_REQUEST_INDICES, "invalid-index")),
-                span ->
-                    span.hasName("callback")
-                        .hasKind(SpanKind.INTERNAL)
-                        .hasParent(trace.getSpan(0))));
-  }
-
-  @Test
-  void elasticsearchGet() {
-    String indexName = "test-index";
-    String indexType = "test-type";
-    String id = "1";
-
-    CreateIndexResponse indexResult = client.admin().indices().prepareCreate(indexName).get();
-    assertThat(indexResult.isAcknowledged()).isTrue();
-
-    GetResponse emptyResult = client.prepareGet(indexName, indexType, id).get();
-    assertThat(emptyResult.isExists()).isFalse();
-    assertThat(emptyResult.getId()).isEqualTo(id);
-    assertThat(emptyResult.getType()).isEqualTo(indexType);
-    assertThat(emptyResult.getIndex()).isEqualTo(indexName);
-
-    IndexResponse createResult =
-        client.prepareIndex(indexName, indexType, id).setSource(Collections.emptyMap()).get();
-    assertThat(createResult.getId()).isEqualTo(id);
-    assertThat(createResult.getType()).isEqualTo(indexType);
-    assertThat(createResult.getIndex()).isEqualTo(indexName);
-    assertThat(createResult.status().getStatus()).isEqualTo(201);
-    cleanup.deferCleanup(() -> client.admin().indices().prepareDelete(indexName).get());
-
-    GetResponse result = client.prepareGet(indexName, indexType, id).get();
-    assertThat(result.isExists()).isTrue();
-    assertThat(result.getId()).isEqualTo(id);
-    assertThat(result.getType()).isEqualTo(indexType);
-    assertThat(result.getIndex()).isEqualTo(indexName);
-
-    // PutMappingAction and IndexAction run in separate threads so their order can vary
-    testing.waitAndAssertSortedTraces(
-        orderByRootSpanName("CreateIndexAction", "PutMappingAction", "IndexAction", "GetAction"),
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
-                    span.hasName("CreateIndexAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasNoParent()
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_ADDRESS,
-                                tcpPublishAddress.getAddress()),
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_PORT, tcpPublishAddress.getPort()),
-                            equalTo(
-                                DbIncubatingAttributes.DB_SYSTEM,
-                                DbIncubatingAttributes.DbSystemValues.ELASTICSEARCH),
-                            equalTo(DbIncubatingAttributes.DB_OPERATION, "CreateIndexAction"),
-                            equalTo(ELASTICSEARCH_ACTION, "CreateIndexAction"),
-                            equalTo(ELASTICSEARCH_REQUEST, "CreateIndexRequest"),
-                            equalTo(ELASTICSEARCH_REQUEST_INDICES, indexName))),
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
-                    span.hasName("PutMappingAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasNoParent()
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(
-                                DbIncubatingAttributes.DB_SYSTEM,
-                                DbIncubatingAttributes.DbSystemValues.ELASTICSEARCH),
-                            equalTo(DbIncubatingAttributes.DB_OPERATION, "PutMappingAction"),
-                            equalTo(ELASTICSEARCH_ACTION, "PutMappingAction"),
-                            equalTo(ELASTICSEARCH_REQUEST, "PutMappingRequest"))),
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
-                    span.hasName("IndexAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasNoParent()
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_ADDRESS,
-                                tcpPublishAddress.getAddress()),
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_PORT, tcpPublishAddress.getPort()),
-                            equalTo(
-                                DbIncubatingAttributes.DB_SYSTEM,
-                                DbIncubatingAttributes.DbSystemValues.ELASTICSEARCH),
-                            equalTo(DbIncubatingAttributes.DB_OPERATION, "IndexAction"),
-                            equalTo(ELASTICSEARCH_ACTION, "IndexAction"),
-                            equalTo(ELASTICSEARCH_REQUEST, "IndexRequest"),
-                            equalTo(ELASTICSEARCH_REQUEST_INDICES, indexName),
-                            equalTo(
-                                AttributeKey.stringKey("elasticsearch.request.write.type"),
-                                indexType),
-                            equalTo(
-                                AttributeKey.longKey("elasticsearch.request.write.version"), -3),
-                            equalTo(AttributeKey.longKey("elasticsearch.response.status"), 201),
-                            equalTo(
-                                AttributeKey.longKey("elasticsearch.shard.replication.total"), 2),
-                            equalTo(
-                                AttributeKey.longKey("elasticsearch.shard.replication.successful"),
-                                1),
-                            equalTo(
-                                AttributeKey.longKey("elasticsearch.shard.replication.failed"),
-                                0))),
-        // moved here by sorting, chronologically happens before PutMappingAction
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
-                    span.hasName("GetAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasNoParent()
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_ADDRESS,
-                                tcpPublishAddress.getAddress()),
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_PORT, tcpPublishAddress.getPort()),
-                            equalTo(
-                                DbIncubatingAttributes.DB_SYSTEM,
-                                DbIncubatingAttributes.DbSystemValues.ELASTICSEARCH),
-                            equalTo(DbIncubatingAttributes.DB_OPERATION, "GetAction"),
-                            equalTo(ELASTICSEARCH_ACTION, "GetAction"),
-                            equalTo(ELASTICSEARCH_REQUEST, "GetRequest"),
-                            equalTo(ELASTICSEARCH_REQUEST_INDICES, indexName),
-                            equalTo(ELASTICSEARCH_TYPE, indexType),
-                            equalTo(ELASTICSEARCH_ID, id),
-                            equalTo(ELASTICSEARCH_VERSION, -1))),
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
-                    span.hasName("GetAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasNoParent()
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_ADDRESS,
-                                tcpPublishAddress.getAddress()),
-                            equalTo(
-                                NetworkAttributes.NETWORK_PEER_PORT, tcpPublishAddress.getPort()),
-                            equalTo(
-                                DbIncubatingAttributes.DB_SYSTEM,
-                                DbIncubatingAttributes.DbSystemValues.ELASTICSEARCH),
-                            equalTo(DbIncubatingAttributes.DB_OPERATION, "GetAction"),
-                            equalTo(ELASTICSEARCH_ACTION, "GetAction"),
-                            equalTo(ELASTICSEARCH_REQUEST, "GetRequest"),
-                            equalTo(ELASTICSEARCH_REQUEST_INDICES, indexName),
-                            equalTo(ELASTICSEARCH_TYPE, indexType),
-                            equalTo(ELASTICSEARCH_ID, id),
-                            equalTo(ELASTICSEARCH_VERSION, 1))));
+  @Override
+  protected int getPort() {
+    return tcpPublishAddress.getPort();
   }
 }
